SFTPClient API Documentation

Base.Filesystem.cd — Method

cd(sftp::SFTP, dir::AbstractString)

Change to dir in the uri of the sftp client.

Base.Filesystem.filemode — Method

Base.isdir(st::SFTPStatStruct) -> UInt

Return the filemode in the SFTPStatStruct.

Base.Filesystem.isdir — Method

isdir(st::SFTPStatStruct) -> Bool

Analyse the SFTPStatStruct and return true for a directory, false otherwise.

Base.Filesystem.isfile — Method

isfile(st::SFTPStatStruct) -> Bool

Analyse the SFTPStatStruct and return true for a file, false otherwise.

Base.Filesystem.islink — Method

islink(st::SFTPStatStruct) -> Bool

Analyse the SFTPStatStruct and return true for a symbolic link, false otherwise.

Base.Filesystem.mkdir — Method

mkdir(sftp::SFTP, dir::AbstractString)

Create a directory dir in the uri of the sftp client.

Base.Filesystem.mv — Method

mv(
    sftp::SFTP,
    old_name::AbstractString,
    new_name::AbstractString;
)

Move, i.e. rename, the file from old_name to new_name in the uri of the sftp client.

Base.Filesystem.pwd — Method

pwd(sftp::SFTP) -> String

Return the current URI path of the SFTP client.

Base.Filesystem.readdir — Function

readdir(sftp::SFTP, join::Bool = false, sort::Bool = true)

Reads the current directory. Returns a vector of Strings just like the regular readdir function.

Base.Filesystem.rm — Method

rm(sftp::SFTP, file::AbstractString)

Remove (delete) the file in the uri of the sftp client.

Base.Filesystem.splitdir — Function

splitdir(sftp::SFTP, path::AbstractString=".") -> Tuple{URI,String}

Join the path with the path of the URI in sftp and then split it into the directory name and base name. Return a Tuple of URI with the split path and a String with the base name.

Base.Filesystem.walkdir — Function

walkdir(
    sftp::SFTP,
    root::AbstractString=".";
    topdown::Bool=true,
    follow_symlinks::Bool=false,
    sort::Bool=true
) -> Channel{Tuple{String,Vector{String},Vector{String}}}

Return an iterator that walks the directory tree of the given root of the sftp client. If the root is ommitted, the current URI path of the sftp client is used. The iterator returns a tuple containing (rootpath, dirs, files). The iterator starts at the root unless topdown is set to false. If follow_symlinks is set to true, the sources of symlinks are listed rather than the symlink itself as file. If sort is set to true, the files and directories are listed alphabetically.

Examples

for (root, dirs, files) in walkdir(sftp, "/")
    println("Directories in $root")
    for dir in dirs
        println(joinpath(root, dir)) # path to directories
    end
    println("Files in $root")
    for file in files
        println(joinpath(root, file)) # path to files
    end
end

Base.download — Function

download(
    sftp::SFTP,
    filename::AbstractString,
    output::String = ""
) -> String

Download a file from the sftp server. The specified filename may include a path on the remote server, which is ignored on the local system.

The file can be downloaded and saved directly to a variable or it can be saved to a file in the output directory.

Example

sftp = SFTP("sftp://test.rebex.net/pub/example/", "demo", "password")
files=readdir(sftp)
download_dir="/tmp"
SFTPClient.download.(sftp, files, download_dir)
````

You can also use it like this:

julia df=DataFrame(CSV.File(SFTPClient.download(sftp, "/mydir/test.csv"))) ```

Base.isequal — Method

isequal(a::SFTPStatStruct, b::SFTPStatStruct) -> Bool

Comares equality between the description (desc fields) of two SFTPStatStruct objects and returns true for equality, otherwise false.

Base.isless — Method

isless(a::SFTPStatStruct, b::SFTPStatStruct) -> Bool

Comares the descriptions (desc fields) of two SFTPStatStruct objects and returns true, if a is lower than b, otherwise false.

Base.stat — Function

stat(sftp::SFTP, path::AbstractString=".") -> SFTPStatStruct

Return the stat data for path on the sftp server.

Note: This returns only stat data for one object, but stat data for all objects in the same folder is obtained internally. If you need stat data for more than object in the same folder, use statscan for better performance and reduced connections to the server.

SFTPClient.change_uripath — Method

change_uripath(uri::URI, path::AbstractString; isfile::Bool=false) -> URI

Return an updated uri struct with the given path. Set isfile to true, if the path is a file to omit the trailing slash.

SFTPClient.check_and_create_fingerprint — Method

check_and_create_fingerprint(host::AbstractString) -> Nothing

Check for host in known_hosts.

SFTPClient.create_fingerprint — Method

create_fingerprint(host::AbstractString) -> Nothing

Create a new entry in known_hosts for host.

SFTPClient.findbase — Method

findbase(stats::Vector{SFTPStatStruct}, base::AbstractString, path::AbstractString) -> Int

Return the index of base in stats or throw and PathNotFoundError, if base is not found.

SFTPClient.ftp_command — Method

ftp_command(sftp::SFTP, command::String)

Execute the command on the sftp server.

SFTPClient.linkerror — Method

linkerror(link::String) -> Nothing

Show an error for an anticipated link format.

SFTPClient.parse_date — Method

parse_date(month::AbstractString, day::AbstractString, year_or_time::AbstractString) -> Float64

From the abbreviated month name, the day and the year_or_time all given as String, return a unix timestamp.

SFTPClient.parse_mode — Method

parse_mode(s::AbstractString) -> UInt

From the AbstractString s, parse the file mode octal number and return as UInt.

SFTPClient.reset_easy_hook — Method

reset_easy_hook(sftp::SFTP) -> Nothing

Reset curl easy options to standard as defined by the sftp client.

SFTPClient.rmdir — Method

rmdir(sftp::SFTP, dir::AbstractString)

Remove (delete) the directory dir in the uri of the sftp client.

SFTPClient.set_stdopt — Method

set_stdopt(sftp::SFTP, easy::Easy) -> Nothing

Set defaults for a number of curl easy options as defined by the sftp client.

SFTPClient.set_url — Method

set_url(url::URI) -> URI

Ensure URI path with trailing slash.

SFTPClient.statscan — Function

statscan(
    sftp::SFTP,
    path::AbstractString=".";
    sort::Bool=true,
    show_cwd_and_parent::Bool=false
) -> Vector{SFTPStatStruct}

Like stat, but returns a Vector of SFTPStatStruct with filesystem stats for all objects in the given path.

** This should be preferred over stat for performance reasons. **

Note that you can only run this on directories.

By default, the SFTPStatStruct vector is sorted by the descriptions (desc fields). For large folder contents, sort can be set to false to increase performance, if the output order is irrelevant. If show_cwd_and_parent is set to true, the SFTPStatStruct vector includes entries for "." and ".." on position 1 and 2, respectively.

SFTPClient.str2number — Method

str2number(s::AbstractString) -> Int64

Parse the file owner symbols in the string s to the corresponding ownership number.

SFTPClient.symlink_source! — Method

symlink_source!(
    sftp::SFTP,
    link::AbstractString,
    pathobjects::@NamedTuple{dirs::Vector{String},files::Vector{String},scans::Dict{String,Any}},
    follow_symlinks::Bool
) -> Nothing

Analyse the symbolic link on the sftp server and add it to the respective pathobjects list. Save the source of the symlink, if follow_symlinks is set to true, otherwise save symlinks as files.

SFTPClient.unescape_joinpath — Method

unescape_joinpath(sftp::SFTP, path::AbstractString) -> String

Join the path with the URI path in sftp and return the unescaped path. Note, this function should not use URL:s since CURL:s api need spaces

SFTPClient.upload — Function

upload(
    sftp::SFTP,
    file::AbstractString;
    remote_dir::AbstractString=".",
    local_dir::AbstractString="."
) -> Nothing

Upload (put) a file on the server. If file includes a path, this is where it is put on the server. The path may be relative to the current uri path of the sftp server or absolute. On the local system, a path may be specified as last argument.

Examples

upload(sftp, "test.csv", "/tmp")

files=readdir()
upload.(sftp, files)

SFTPClient.PathNotFoundError — Type

PathNotFoundError(path)

The path (file or folder) was not found.

SFTPClient.SFTP — Type

mutable struct SFTP

SFTP manages the connection to the server and stores all relevant connection data.

Constructors

SFTP(url::AbstractString, username::AbstractString, public_key_file::AbstractString, public_key_file::AbstractString; kwargs) -> SFTP
SFTP(url::AbstractString, username::AbstractString, password::AbstractString=""; kwargs) -> SFTP

Construct an SFTP client from the url and either user information or public and private key file.

Arguments

url: The url to connect to, e.g., sftp://mysite.com

username/password: user credentials

public_key_file/public_key_file: authentication certificates

Keyword arguments

create_known_hosts_entry: Automatically create an entry in known hosts

disable_verify_peer: verify the authenticity of the peer's certificate

disable_verify_host: verify the host in the server's TLS certificate

verbose: display a lot of verbose curl information

Important notice

A username must be provided for both methods to work.

Before using the constructor method for certificate authentication, private and public key files must be created and stored in the ~/.ssh folder and on the server, e.g. ~/.ssh/idrsa and ~/.ssh/idrsa.pub. Additionally, the host must be added to the known_hosts file in the ~/.ssh folder.

The correct setup can be tested in the terminal with ssh myuser@mysitewhereIhaveACertificate.com.

Examples

sftp = SFTP("sftp://mysitewhereIhaveACertificate.com", "myuser", "test.pub", "test.pem")
sftp = SFTP("sftp://mysitewhereIhaveACertificate.com", "myuser")
sftp = SFTP("sftp://test.rebex.net", "demo", "password")

SFTPClient.SFTPStatStruct — Type

struct SFTPStatStruct

Hold information for file system objects on a Server.

Fields

desc::String: file or folder description/name

mode::UInt: file system object type (file, folder, etc.)

nlink::Int: number of hard links (contents)

uid::String: numeric user ID of the owner of the file/folder

uid::String: numeric group ID (gid) for the file/folder

size::Int64: file/folder size in Byte

mtime::Float64: modified time

Constructors

SFTPStatStruct(stats::AbstractString) -> SFTPStatStruct

Parse the stats string and return an SFTPStatStruct.

The stats are of the format:

"d--x--x---  151 ftp      ftp          8192 Dec  2  2023 .."